Add STM32 external loader support
This patch introduces st-extflash tool.
The goal of this tool is to program external memories embedded on STM32 based
boards, which was only possible with ST's ST-Link Utility for Windows until
now.

This tool reuses the loaders (.stldr files) provided with ST-Link Utility,
and supports erase, read and write (with or without verification)
features.

The tool has been tested with QSPI NOR of the STM32F469i-Discovery board,
and with NOR/QSPI NOR of the STM32746g-EVAL2 board.

For usage information, please try "st-exflash --help".

Signed-off-by: Maxime Coquelin <mcoquelin.stm32@gmail.com>